Healthbar works

+++ Assets/_SCRIPTS/PlayerHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
     public HeroKnight Player;
 
     void Start()
